Update page to be Origin focused project page
This commit includes the updated index.html and all the supporting
asset files for the first draft of the page.

Background: Since we are automatically building the Origin
documenation as a subdirectory, we want a corresponding landing
page to serve as an open source project page just like any other open
source project hosted on GitHub. The purpose of the page is two-fold:
to point to the documentation and to give a overview of the project.